backend/pyproject.toml
[project]
name = "trustflow"
version = "0.1.0"
description = "TrustFlow - Escrow payment platform with ETH guarantee"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Configuration & Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Blockchain
    "web3>=7.6.0",
    # Auth (JWT validation)
    "python-jose[cryptography]>=3.3.0",
    # HTTP Client
    "httpx>=0.28.0",
    "google-auth>=2.48.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    
    # Code Quality
    "ruff>=0.8.0",
    "black>=24.10.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "src"]
python_files = ["test_*.py", "*_test.py"]

[tool.black]
line-length = 88
target-version = ["py312"]
